separate field collection data from logic
